[flang][runtime] Decouple scalar output APIs from descriptors
For testing purposes the implementations of the output APIs
like OutputInteger32 have been simply constructing descriptors
and executing the operation through the general DescriptorIO
template.  This patch decouples those APIs from that mechanism
so that programs using simple "PRINT *" statements for output
can link to a smaller portion of the I/O runtime support library.
(This is the only form of I/O accepted in GPU device code by
previous CUDA Fortran and Fortran OpenACC compilers.)
